Application/requirements.txt
streamlit=1.41.1
pillow=11.0.0